Moved code to namespaces and updated enemies with hp system and extra animations
