+++ Assets/Characters/Flying eye/Animations/Death.anim.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: e7323796b96a90d429ad89c9a5003ab5
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 7400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
