+++ Assets/Characters/Flying eye/Animations/TakeHit.anim.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: a099e7586a803d541ad5ce5de1c6fe8a
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 7400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
