+++ Assets/Characters/Goblin/Animations/Death.anim.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 89f23d56faabebf4e9eca72cc98264c4
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 7400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
