+++ Assets/Characters/Goblin/Animations/Goblin.controller.meta
