+++ Assets/Characters/Goblin/Animations/TakeHit.anim.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 10cd4d2fef9d9014c82a688944c63aca
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 7400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
