+++ Assets/Characters/Mushroom/Animations/Death.anim.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: ad1dcb1184c008348ad8aea96a369cfc
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 7400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
