+++ Assets/Characters/Mushroom/Animations/TakeHit.anim.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: a07895a63f227994d96b15de70cdfd4e
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 7400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
