+++ Assets/Characters/Skeleton/Animations/Death.anim.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 043b89724e4a3b447af72653b72f64c8
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 7400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
