+++ Assets/Characters/Skeleton/Animations/TakeHit.anim.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 33107592e9090154a9acd3353313916d
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 7400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
