+++ Assets/Prefabs/Enemies/Bringer Of Death.prefab
@@ -13,6 +13,7 @@ GameObject:
   - component: {fileID: 9148402442353990094}
   - component: {fileID: 2634237087749244889}
   - component: {fileID: 7955565140460621189}
+  - component: {fileID: 303159700903167475}
   m_Layer: 0
   m_Name: Bringer Of Death
   m_TagString: Untagged
@@ -155,3 +156,15 @@ BoxCollider2D:
   serializedVersion: 2
   m_Size: {x: 0.4, y: 0.6}
   m_EdgeRadius: 0
+--- !u!114 &303159700903167475
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 9148402442353990093}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 4baee0ea5421f9346b1277a500cdbc73, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier:
