+++ Assets/Prefabs/Enemies/Flying Eye.prefab
@@ -13,6 +13,7 @@ GameObject:
   - component: {fileID: 7249707510761590529}
   - component: {fileID: 8444605326104604760}
   - component: {fileID: 9212963108982491359}
+  - component: {fileID: -795928965890314731}
   m_Layer: 0
   m_Name: Flying Eye
   m_TagString: Untagged
@@ -155,3 +156,15 @@ BoxCollider2D:
   serializedVersion: 2
   m_Size: {x: 0.4, y: 0.6}
   m_EdgeRadius: 0
+--- !u!114 &-795928965890314731
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 2575639233869638719}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 4baee0ea5421f9346b1277a500cdbc73, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier:
