+++ Assets/Prefabs/Enemies/Mushroom.prefab
@@ -13,6 +13,7 @@ GameObject:
   - component: {fileID: 5822868078199210971}
   - component: {fileID: 4297946461705690994}
   - component: {fileID: 57242523094245262}
+  - component: {fileID: -6150724485947031455}
   m_Layer: 0
   m_Name: Mushroom
   m_TagString: Untagged
@@ -155,3 +156,15 @@ BoxCollider2D:
   serializedVersion: 2
   m_Size: {x: 0.2, y: 0.4}
   m_EdgeRadius: 0
+--- !u!114 &-6150724485947031455
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 5822868078199210968}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 4baee0ea5421f9346b1277a500cdbc73, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier:
