+++ Assets/Prefabs/Enemies/Skeleton.prefab
@@ -13,6 +13,7 @@ GameObject:
   - component: {fileID: 7450140168799887511}
   - component: {fileID: 6650660553369781682}
   - component: {fileID: 1894763928854331813}
+  - component: {fileID: -6963178884822222663}
   m_Layer: 0
   m_Name: Skeleton
   m_TagString: Untagged
@@ -155,3 +156,15 @@ BoxCollider2D:
   serializedVersion: 2
   m_Size: {x: 0.4, y: 0.5}
   m_EdgeRadius: 0
+--- !u!114 &-6963178884822222663
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 7450140168799887496}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 4baee0ea5421f9346b1277a500cdbc73, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier:
