+++ Assets/Scripts/Characters.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: f1bccecf99076f44496c014e84d8ba7e
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
