+++ Assets/Scripts/Characters/Enemies.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 4a8ff1738338e9a4ba1a2b8f512fc59b
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
