+++ Assets/Scripts/Characters/Enemies/Enemy.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 79f4bd0f672f9e84da175dfb56355150
+guid: 4baee0ea5421f9346b1277a500cdbc73
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
