+++ Assets/Scripts/Characters/Player.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 9483ec0164f99f243a6646f987fd61a4
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
