+++ Assets/Scripts/Characters/Player/AoeAttack.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: d926964c3df1b88469a9db02c35ea14c
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
