+++ Assets/Scripts/Dungeon/Areas/Corridors/CorridorVisualizer.cs
@@ -1,3 +1,4 @@
+using Assets.Scripts.Extensions;
 using System.Collections.Generic;
 using System.Linq;
 using UnityEditor;
