+++ Assets/Scripts/Dungeon/Areas/Rooms/Room.cs
@@ -1,3 +1,4 @@
+using Assets.Scripts.Extensions;
 using UnityEngine;
 
 namespace Assets.Scripts.Dungeon.Areas.Rooms
