+++ Assets/Scripts/Dungeon/Areas/Rooms/RoomVisualizer.cs
@@ -1,3 +1,4 @@
+using Assets.Scripts.Extensions;
 using System.Collections.Generic;
 using System.Linq;
 using UnityEngine;
