+++ Assets/Scripts/Extensions/PositionEnumeratorExtensions.cs
@@ -2,15 +2,18 @@ using System.Collections.Generic;
 using UnityEngine;
 using static UnityEngine.RectInt;
 
-public static class PositionEnumeratorExtensions
+namespace Assets.Scripts.Extensions
 {
-    public static List<Vector2Int> ToVector2Int(this PositionEnumerator positionEnumerator)
+    public static class PositionEnumeratorExtensions
     {
-        var vector2IntList = new List<Vector2Int>();
-        foreach (var item in positionEnumerator)
+        public static List<Vector2Int> ToVector2Int(this PositionEnumerator positionEnumerator)
         {
-            vector2IntList.Add(item);
+            var vector2IntList = new List<Vector2Int>();
+            foreach (var item in positionEnumerator)
+            {
+                vector2IntList.Add(item);
+            }
+            return vector2IntList;
         }
-        return vector2IntList;
     }
 }
