--- Assets/Scripts/MeleeAttack.cs
@@ -1,28 +1,0 @@
-using System.Collections;
-using System.Collections.Generic;
-using UnityEngine;
-
-public class MeleeAttack : MonoBehaviour
-{
-    /*Collider2D tentacleCollider;
-
-    private void Start()
-    {
-        tentacleCollider = GetComponent<Collider2D>();
-    }
-
-    public void AttackRight
-    {
-
-    }
-
-    public void AttackLeft
-    {
-
-    }
-
-    public void StopAttack
-    {
-
-    }*/
-}
